update edge ui
